Use argh for argument parsing (fixes #12)
