+++ Cargo.toml
@@ -11,16 +11,17 @@ license = "MIT"
 readme = "README.md"
 
 [dependencies]
+anyhow = "1.0.57"
+argh = "0.1.9"
+cargo_metadata = "0.14.2"
 cargo_toml = "0.13.0"
 grep = "0.2.8"
 log = "0.4.16"
 pretty_env_logger = "0.4.0"
-walkdir = "2.3.2"
-anyhow = "1.0.57"
 rayon = "1.5.2"
-cargo_metadata = "0.14.2"
 serde = "1.0.136"
 toml_edit = { version = "0.14.3", features = ["easy", "serde"] }
+walkdir = "2.3.2"
 
 # Uncomment this for profiling.
 #[profile.release]
